refactor infection modeling
